+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Changed
 - Target handlebars.c version to 0.6.0
 
+### Changed
+- Target handlebars.c version to 0.6.0
+
 ## [0.7.4] - 2016-09-21
 
 ### Added